+++ udata/core/spam/models.py
@@ -97,7 +97,7 @@ class SpamMixin(object):
 
             # Language detection is not working well with texts of a few words.
             if SpamMixin.allowed_langs() and len(text) > SpamMixin.minimum_string_length_for_lang_check():
-                lang = detect(text)
+                lang = detect(text.lower())
                 if lang not in SpamMixin.allowed_langs():
                     self.spam.status = POTENTIAL_SPAM
                     self._report(text=text, breadcrumb=breadcrumb, reason=f"not allowed language \"{lang}\"")
@@ -143,11 +143,8 @@ class SpamMixin(object):
     def embeds_to_check_for_spam(self):
         return []
 
-    def spam_report_title(self):
-        return type(self).__name__
-
-    def spam_report_link(self):
-        return None
+    def spam_report_message(self):
+        return f"Spam potentiel sur {type(self).__name__}"
 
     def _report(self, text, breadcrumb, reason):
         base_model = breadcrumb[0]
@@ -158,22 +155,14 @@ class SpamMixin(object):
             if document != base_model:
                 return
 
-            # Note that all the chain should be a SpamMixin, maybe we could filter out if it's not the case here…
-            title = " → ".join(map(lambda o: o.spam_report_title(), breadcrumb))
+            message = self.spam_report_message(breadcrumb)
 
-            # Select the first link in the embed list (for example message doesn't have a link, so check if discussion
-            # have one)
-            for object in reversed(breadcrumb):
-                link = object.spam_report_link()
-                if link:
-                    break
-
-            on_new_potential_spam.send(self, title=title, link=link, text=text, reason=reason)
+            on_new_potential_spam.send(self, message=message, text=text, reason=reason)
 
             # We clean the listener here. Not sure if it's necessary either.
             signals.post_save.disconnect(report_after_save)
 
-        # For things like `spam_report_link` we often need the ID of the document so we
+        # For `spam_report_message` we often need the ID of the document so we
         # must report after saving to have the ID available.
         # By default the signal is weak so it is dropped at the end of this function and it's
         # never called. We disconnect the signal in `report_after_save` to avoid leaks.